Split EGD/PRNGD interface into its own file.
This will allow us to use it when building --without-openssl.
